+++ stencil-workspace/src/components/icons/IconProps.tsx
@@ -1,0 +1,9 @@
+interface IconProps {
+  color?: string;
+  height?: string;
+  onClick?: (event: Event) => void;
+  size?: string;
+  width?: string;
+}
+
+export default IconProps;